10-7959c | Reorganize tests and utils file structures
