--- src/applications/ivc-champva/10-7959C/helpers/utilities.js
@@ -1,7 +1,0 @@
-export function isRequiredFile(formContext, requiredFiles) {
-  return Object.keys(formContext?.schema?.properties || {}).filter(v =>
-    Object.keys(requiredFiles).includes(v),
-  ).length >= 1
-    ? '(Required)'
-    : '(Optional)';
-}
