--- src/applications/ivc-champva/10-7959C/tests/containers/IntroductionPage.unit.spec.jsx
@@ -1,6 +1,0 @@
-import formConfig from '../../config/form';
-import IntroductionPage from '../../containers/IntroductionPage';
-
-import { TestIntroPageRenderAndStartLink } from '../../../shared/tests/components/IntroductionPage.unit.spec';
-
-TestIntroPageRenderAndStartLink(formConfig, IntroductionPage);
