+++ src/applications/ivc-champva/10-7959C/tests/unit/config/form.unit.spec.jsx
@@ -1,9 +1,9 @@
 import { expect } from 'chai';
-import { testNumberOfWebComponentFields } from '../../../shared/tests/pages/pageTests.spec';
-import formConfig from '../../config/form';
-import mockData from '../e2e/fixtures/data/test-data.json';
-import maxData from '../e2e/fixtures/data/maximal-test.json';
-import FileFieldWrapped from '../../components/FileUploadWrapper';
+import { testNumberOfWebComponentFields } from '../../../../shared/tests/pages/pageTests.spec';
+import formConfig from '../../../config/form';
+import mockData from '../../e2e/fixtures/data/test-data.json';
+import maxData from '../../e2e/fixtures/data/maximal-test.json';
+import FileFieldWrapped from '../../../components/FileUploadWrapper';
 
 // Call the depends() function for any page that relies on it
 describe('dependent page logic', () => {
