+++ src/applications/ivc-champva/10-7959C/tests/unit/config/migrations.unit.spec.jsx
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { migrateCardUploadKeys } from '../../config/migrations';
+import { migrateCardUploadKeys } from '../../../config/migrations';
 
 const EXAMPLE_SIP_METADATA = {
   version: 0,
