+++ src/applications/ivc-champva/10-7959C/tests/unit/config/transformers.unit.spec.jsx
@@ -1,9 +1,8 @@
 import { expect } from 'chai';
-import formConfig from '../../config/form';
-import mockData from '../e2e/fixtures/data/test-data.json';
-
-import transformForSubmit from '../../config/submitTransformer';
-import { prefillTransformer } from '../../config/prefillTransformer';
+import formConfig from '../../../config/form';
+import mockData from '../../e2e/fixtures/data/test-data.json';
+import transformForSubmit from '../../../config/submitTransformer';
+import { prefillTransformer } from '../../../config/prefillTransformer';
 
 describe('Prefill transformer', () => {
   it('should return pages, formData, and metadata', () => {
