+++ src/applications/ivc-champva/10-7959C/tests/unit/containers/IntroductionPage.unit.spec.jsx
@@ -1,0 +1,5 @@
+import formConfig from '../../../config/form';
+import IntroductionPage from '../../../containers/IntroductionPage';
+import { TestIntroPageRenderAndStartLink } from '../../../../shared/tests/components/IntroductionPage.unit.spec';
+
+TestIntroPageRenderAndStartLink(formConfig, IntroductionPage);
