+++ src/applications/ivc-champva/10-7959C/utils/titles.js
@@ -1,5 +1,5 @@
 import { arrayBuilderItemSubsequentPageTitleUI } from 'platform/forms-system/src/js/web-component-patterns';
-import { replaceStrValues } from '../helpers/formatting';
+import { replaceStrValues } from './helpers/formatting';
 
 /**
  * Creates a dynamic Health Insurance page title with provider name.
